useCaptureReportSnapshot
